无缓存时删除crash 替换icon

+++ SwiftQRCodeScan/HistoryDataCache.swift
@@ -92,7 +92,9 @@ class HistoryDataCache {
     private func removeCacheFile() {
         weak var weakSelf = self
         DispatchQueue.main.async {
-            try! FileManager.default.removeItem(atPath: (weakSelf?.fileName())!)
+            if FileManager.default.fileExists(atPath: (weakSelf?.fileName())!) {
+                try! FileManager.default.removeItem(atPath: (weakSelf?.fileName())!)
+            }
         }
     }
 }
